last styling

+++ routes/index.js
@@ -7,34 +7,39 @@ const Records = require('../models/Records.model');
 const { response } = require('../app');
 /* GET home page */
 router.get('/', (req, res, next) => {
-  res.render('index');
+  res.render('index', {userInSession: req.user});
 });
+
+
 
 // Get the games page
 router.get('/games',(req,res,next) => {
+  console.log(req.user);
+  if(!req.user){
+    res.redirect('/')
+  }
   axios
   .get('https://api.rawg.io/api/games')
   .then ((response) =>{
-    // console.log("Response from the API: ", response.data);
-    //res.render('games', {games: response.data.results})
+    
     let results = response.data.results;
-    //const {name,background_image,suggestions_count,metacritic,released,platforms} = response.data.results;
+    
     Games.create(results)
-    .then((resultsFromDB) => {
-
-      // console.log(`created ${resultsFromDB.length} games`);
-      res.render('games', {games: resultsFromDB})
-  })
-   .catch((error) =>{
-  if(error.code === 11000){
-    Games.find()
       .then((resultsFromDB) => {
-        res.render('games', {games: resultsFromDB})
+
+          // console.log(`created ${resultsFromDB.length} games`);
+          res.render('games', {games: resultsFromDB, userInSession: req.user})
       })
-    return
-  }
-  next(error)
-})
+      .catch((error) =>{
+          if(error.code === 11000){
+            Games.find()
+              .then((resultsFromDB) => {
+                res.render('games', {games: resultsFromDB,  userInSession: req.user})
+              })
+            return
+          }
+          next(error)
+      })
 
   })
   .catch((error) => {
@@ -50,10 +55,16 @@ router.get('/hof', (req, res, next) => {
   res.render('hof');
 });
 router.get('/streams', (req, res, next) => {
-  res.render('streams');
+  if(!req.user){
+    res.redirect('/')
+  }
+  res.render('streams', {userInSession: req.user});
 });
 
-router.get('/games/:id', (req,res) =>{
+router.get('/games/:id', (req,res, next) =>{
+  if(!req.user){
+    res.redirect('/')
+  }
   console.log(req.params);
   const { id } = req.params
 
@@ -61,27 +72,30 @@ router.get('/games/:id', (req,res) =>{
     games: mongoose.Types.ObjectId(id)
   })
   .then((response) => {
-    console.log('RECORDS:', response)
-    res.render('records', {
-      game_id: id,
-      records: response
-    });
+    // console.log('RECORDS:', response)
+    Games.findById(id).then((foundGame) => {
+      // console.log('FOUND:', foundGame.name)
+      res.render('records', {
+        game_id: id,
+        records: response,
+        game: foundGame.name,
+        userInSession: req.user
+      });
+    }).catch((error) => next(error))
   })
     .catch(() => {
       res.send("Nope, check if everything is working")
     })
 
-  // Games.findById(id)
-  // .then((response) => {
-  //   console.log('GAME', response)
-  //   res.render('records', response);
-  // })
-  //   .catch(() => {
-  //     res.send("Nope, check if everything is working")
-  //   })
+
 })
+
+
  
 router.get('/games/:id/records/create', (req,res) => {
+  if(!req.user){
+    res.redirect('/')
+  }
   const { id } = req.params
  res.render('create-record.hbs', {id})
 });
@@ -92,7 +106,9 @@ router.post('/games/:id/records/create', (req,res) => {
   const { id } = req.params
   const _id = mongoose.Types.ObjectId(id);
 
-  Records.create({name,time,picture,video, games:[ _id ]})
+  
+
+  Records.create({name,time,picture,video, games:[ _id ], users:[req.user.id]})
   .then(() => {
     const { id } = req.params
     res.redirect(`/games/${id}`)
@@ -104,8 +120,25 @@ router.post('/games/:id/records/create', (req,res) => {
 
  router.get('/games/:id/records/edit', (req,res) => {
   const {id} = req.params
+  if(!req.user){
+    res.redirect('/');
+    return;
+  }
+  
+  
   Records.findById(id)
   .then((response) => {
+    console.log(response);
+    
+    if (!new mongoose.Types.ObjectId(req.user._id).equals(response.users[0])){
+      return res.redirect('/')
+      }
+    // if(!req.user || req.user._id !== response.users[0]){
+    //   console.log(response); // is not author
+    //   res.redirect('/');
+    //   return
+    // }
+    
     res.render('update-form', response);
   })
     .catch(() => {
@@ -132,13 +165,29 @@ router.post('/games/:id/records/edit', (req,res) => {
 router.post('/games/:game_id/records/:record_id/delete',(req,res,next) => {
   const {game_id, record_id} = req.params
   //const { name, time, picture, video} = req.body
-  Records.findByIdAndDelete(record_id)
-  .then((doc) => {
-    res.redirect('/games/'+game_id)
+
+
+  Records.findById(record_id)
+  .then((response) => {
+
+    if (!new mongoose.Types.ObjectId(req.user._id).equals(response.users[0])){
+      return res.redirect('/')
+      }
+    
+    Records.findByIdAndDelete(record_id)
+      .then((doc) => {
+        res.redirect('/games/'+game_id)
+      })
+      .catch((response) => {
+        res.render('update-form',{response})
+      })
   })
-  .catch((response) => {
-    res.render('update-form',{response})
+  .catch(() => {
+    res.send("Nope, check if everything is working")
   })
+  
+  
+
 });
 module.exports = router;
